# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PSA Backend - All Projects",
    "version": "1.0.0",
    "description": "Every endpoint of every project backend in this repository. Operations are tagged and prefixed with the project they belong to."
  },
  "paths": {
...
    "/membership-analytics/api/v1/nominated-post/bootstrap": {
      "get": {
        "tags": [
          "Membership Analytics - Nominated Post"
        ],
        "summary": "[Membership Analytics] Nominated Bootstrap",
        "operationId": "Membership_Analytics__nominated_bootstrap_api_v1_nominated_post_bootstrap_get",
        "parameters": [
          {
            "name": "enrollmentId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "title": "Enrollmentid"
            }
          },
          {
            "name": "boardLevelId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Boardlevelid"
            }
          },
          {
            "name": "locationValue",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Locationvalue"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Membership_Analytics__app__schemas__nominated_schema__ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Membership_Analytics__HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/membership-analytics/api/v1/nominated-post/locations": {
      "get": {
        "tags": [
          "Membership Analytics - Nominated Post"
        ],
        "summary": "[Membership Analytics] Locations",
        "operationId": "Membership_Analytics__locations_api_v1_nominated_post_locations_get",
        "parameters": [
          {
            "name": "boardLevelId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Boardlevelid"
            }
          },
          {
            "name": "stateId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Stateid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Membership_Analytics__app__schemas__nominated_schema__ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Membership_Analytics__HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/membership-analytics/api/v1/nominated-post/departments": {
      "get": {
        "tags": [
          "Membership Analytics - Nominated Post"
        ],
        "summary": "[Membership Analytics] Departments",
        "description": "Returns all departments for the selected level/location with totalSeats,\nfilledSeats, and unfilledSeats per department.\n\nExample:\n/api/v1/nominated-post/departments?enrollmentId=2&boardLevelId=2&locationValue=1",
        "operationId": "Membership_Analytics__departments_api_v1_nominated_post_departments_get",
        "parameters": [
          {
            "name": "enrollmentId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "title": "Enrollmentid"
            }
          },
          {
            "name": "boardLevelId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Boardlevelid"
            }
          },
          {
            "name": "locationValue",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Locationvalue"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Membership_Analytics__app__schemas__nominated_schema__ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Membership_Analytics__HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/membership-analytics/api/v1/nominated-post/boards": {
      "get": {
        "tags": [
          "Membership Analytics - Nominated Post"
        ],
        "summary": "[Membership Analytics] Boards",
        "description": "Returns boards with totalSeats, filledSeats, and unfilledSeats. When departmentId\nis supplied, scopes to that department. When omitted, returns ALL boards for the\nlevel/location, each carrying its owning departmentId/departmentName.",
        "operationId": "Membership_Analytics__boards_api_v1_nominated_post_boards_get",
        "parameters": [
          {
            "name": "enrollmentId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "title": "Enrollmentid"
            }
          },
          {
            "name": "boardLevelId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Boardlevelid"
            }
          },
          {
            "name": "locationValue",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Locationvalue"
            }
          },
          {
            "name": "departmentId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Departmentid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Membership_Analytics__app__schemas__nominated_schema__ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Membership_Analytics__HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/membership-analytics/api/v1/nominated-post/positions": {
      "get": {
        "tags": [
          "Membership Analytics - Nominated Post"
        ],
        "summary": "[Membership Analytics] Positions",
        "description": "Returns all positions for the selected department and board with totalSeats,\nfilledSeats, and unfilledSeats, plus nominatedPostMemberId / nominatedPostPositionId.",
        "operationId": "Membership_Analytics__positions_api_v1_nominated_post_positions_get",
        "parameters": [
          {
            "name": "enrollmentId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "title": "Enrollmentid"
            }
          },
          {
            "name": "boardLevelId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Boardlevelid"
            }
          },
          {
            "name": "locationValue",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Locationvalue"
            }
          },
          {
            "name": "departmentId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Departmentid"
            }
          },
          {
            "name": "boardId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Boardid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Membership_Analytics__app__schemas__nominated_schema__ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Membership_Analytics__HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/membership-analytics/api/v1/nominated-post/capacity": {
      "get": {
        "tags": [
          "Membership Analytics - Nominated Post"
        ],
        "summary": "[Membership Analytics] Capacity",
        "operationId": "Membership_Analytics__capacity_api_v1_nominated_post_capacity_get",
        "parameters": [
          {
            "name": "enrollmentId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "title": "Enrollmentid"
            }
          },
          {
            "name": "boardLevelId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Boardlevelid"
            }
          },
          {
            "name": "locationValue",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Locationvalue"
            }
          },
          {
            "name": "departmentId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Departmentid"
            }
          },
          {
            "name": "boardId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Boardid"
            }
          },
          {
            "name": "positionId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Positionid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Membership_Analytics__app__schemas__nominated_schema__ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Membership_Analytics__HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/membership-analytics/api/v1/nominated-post/cadre/search": {
      "get": {
        "tags": [
          "Membership Analytics - Nominated Post"
        ],
        "summary": "[Membership Analytics] Search Cadre",
        "operationId": "Membership_Analytics__search_cadre_api_v1_nominated_post_cadre_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 2,
              "title": "Q"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Membership_Analytics__app__schemas__nominated_schema__ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Membership_Analytics__HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/membership-analytics/api/v1/nominated-post/cache/refresh": {
      "post": {
        "tags": [
          "Membership Analytics - Nominated Post"
        ],
        "summary": "[Membership Analytics] Refresh Cache",
        "operationId": "Membership_Analytics__refresh_cache_api_v1_nominated_post_cache_refresh_post",
        "parameters": [
          {
            "name": "enrollmentId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "title": "Enrollmentid"
            }
          },
          {
            "name": "boardLevelId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Boardlevelid"
            }
          },
          {
            "name": "locationValue",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Locationvalue"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Membership_Analytics__app__schemas__nominated_schema__ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Membership_Analytics__HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Membership_Analytics__HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/Membership_Analytics__ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Membership_Analytics__ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "Membership_Analytics__app__schemas__nominated_schema__ApiResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "title": "Data"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "OK"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ApiResponse"
      },
...
    }
  },
...
}